RF: Handle failed install in pkgtools
