+++ psychopy/tools/pkgtools.py
@@ -22,11 +22,73 @@ __all__ = [
 import subprocess as sp
 from psychopy.preferences import prefs
 from psychopy.localization import _translate
+from psychopy.app.themes import icons, handlers
 import pkg_resources
 import sys
 import os
 import requests
 import wx
+
+
+class InstallErrorDlg(wx.Dialog, handlers.ThemeMixin):
+    """
+    Dialog to display when something fails to install, contains info on what
+    command was tried and what output was received.
+    """
+    def __init__(self, label, caption=_translate("PIP error"), cmd="", stdout="", stderr=""):
+        from psychopy.app.themes import fonts
+        # Initialise
+        wx.Dialog.__init__(
+            self, None,
+            size=(480, 620),
+            title=caption,
+            style=wx.RESIZE_BORDER | wx.CLOSE_BOX | wx.CAPTION
+        )
+        # Setup sizer
+        self.border = wx.BoxSizer(wx.VERTICAL)
+        self.SetSizer(self.border)
+        self.sizer = wx.BoxSizer(wx.VERTICAL)
+        self.border.Add(self.sizer, proportion=1, border=6, flag=wx.ALL | wx.EXPAND)
+        # Create title sizer
+        self.title = wx.BoxSizer(wx.HORIZONTAL)
+        self.sizer.Add(self.title, border=6, flag=wx.ALL | wx.EXPAND)
+        # Create icon
+        self.icon = wx.StaticBitmap(
+            self, size=(32, 32),
+            bitmap=icons.ButtonIcon(stem="stop", size=32).bitmap
+        )
+        self.title.Add(self.icon, border=6, flag=wx.ALL | wx.EXPAND)
+        # Create title
+        self.titleLbl = wx.StaticText(self, label=label)
+        self.titleLbl.SetFont(fonts.appTheme['h3'].obj)
+        self.title.Add(self.titleLbl, proportion=1, border=6, flag=wx.ALL | wx.ALIGN_CENTER_VERTICAL)
+        # Show what we tried
+        self.inLbl = wx.StaticText(self, label=_translate("We tried:"))
+        self.sizer.Add(self.inLbl, border=6, flag=wx.ALL | wx.EXPAND)
+        self.inCtrl = wx.TextCtrl(self, value=cmd, style=wx.TE_READONLY)
+        self.inCtrl.SetBackgroundColour("white")
+        self.inCtrl.SetFont(fonts.appTheme['code'].obj)
+        self.sizer.Add(self.inCtrl, border=6, flag=wx.ALL | wx.EXPAND)
+        # Show what we got
+        self.outLbl = wx.StaticText(self, label=_translate("We got:"))
+        self.sizer.Add(self.outLbl, border=6, flag=wx.ALL | wx.EXPAND)
+        self.outCtrl = wx.TextCtrl(self, value=f"{stdout}\n{stderr}",
+                                   size=(-1, 620), style=wx.TE_READONLY | wx.TE_MULTILINE)
+        self.outCtrl.SetFont(fonts.appTheme['code'].obj)
+        self.sizer.Add(self.outCtrl, proportion=1, border=6, flag=wx.ALL | wx.EXPAND)
+
+        # Make buttons
+        self.btns = self.CreateStdDialogButtonSizer(flags=wx.OK)
+        self.border.Add(self.btns, border=6, flag=wx.ALIGN_RIGHT | wx.ALL)
+
+        self.Layout()
+        self._applyAppTheme()
+
+    def ShowModal(self):
+        # Make error noise
+        wx.Bell()
+        # Show as normal
+        wx.Dialog.ShowModal(self)
 
 
 def getDistributions():
@@ -128,10 +190,25 @@ def installPackage(package, target=None, upgrade=False, forceReinstall=False,
     sys.stdout.write(stdout)
     sys.stderr.write(stderr)
 
-    if stderr:   # any error, return False
-        return False
-
-    return True
+    # Handle errors
+    if output.returncode != 0:
+        # Display output if error
+        cmd = "\n>> pip install" + package + "\n"
+        dlg = InstallErrorDlg(
+            cmd=cmd,
+            stdout=stdout,
+            stderr=stderr,
+            label=_translate("Package {} could not be installed.").format(package)
+        )
+    else:
+        # Display success message if success
+        dlg = wx.MessageDialog(
+            parent=None,
+            caption=_translate("Package installed"),
+            message=_translate("Package {} successfully installed!").format(package),
+            style=wx.ICON_INFORMATION
+        )
+    dlg.ShowModal()
 
 
 def isInstalled(packageName):
